UserServiceTests 실행

+++ src/main/java/com/ll/exam/sbbfinal/answer/AnswerController.java
@@ -28,7 +28,7 @@ public class AnswerController {
 
 
     @PostMapping("/create/{id}")
-    public String detail(Model model, @PathVariable int id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
+    public String detail(Model model, @PathVariable long  id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
         Question question = questionService.getQuestion(id);
 
         if(bindingResult.hasErrors()){
